backend/tests/test_demo_script.py
"""Demo Script Test (T076, T078).

Runs the 5 demo interactions from quickstart.md to verify all functionality works.
"""

import os
import sys
import time
import uuid

import jwt
import requests

# Add parent directories to path for imports
sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))

from dotenv import load_dotenv

load_dotenv()

BACKEND_URL = os.getenv("BACKEND_URL", "http://localhost:8000")
CHATKIT_SECRET_KEY = os.getenv("CHATKIT_SECRET_KEY", "default-secret-key-change-in-production")


def create_chatkit_token(user_id: str) -> str:
    """Create a ChatKit JWT token for testing."""
    now = int(time.time())
    payload = {
        "sub": user_id,
        "exp": now + 3600,
        "iat": now,
        "iss": "todoapp-chatkit",
    }
    return jwt.encode(payload, CHATKIT_SECRET_KEY, algorithm="HS256")


def check_server_health() -> bool:
    """Check if backend server is running."""
    try:
        response = requests.get(f"{BACKEND_URL}/api/v1/health", timeout=5)
        return response.ok
    except requests.exceptions.RequestException:
        return False


def run_demo_script():
    """Run the 5 demo interactions from quickstart.md."""
    print("\n" + "#" * 60)
    print("# Demo Script Test (T076, T078)")
    print("# From: specs/003-ai-todo-chatbot/quickstart.md")
    print("#" * 60)

    # Check server health
    if not check_server_health():
        print("\nERROR: Backend server not available at", BACKEND_URL)
        print("Please start the server with: uv run uvicorn app.main:app --reload")
        return 1

    print(f"\nServer: {BACKEND_URL} [OK]")

    # Create a unique test user for this demo
    user_id = f"demo-user-{uuid.uuid4().hex[:8]}"
    token = create_chatkit_token(user_id)
    print(f"Test User: {user_id}\n")

    def chat(message: str) -> str:
        """Send a chat message and return the response."""
        try:
            response = requests.post(
                f"{BACKEND_URL}/api/chatkit/chat",
                headers={"Authorization": f"Bearer {token}"},
                json={"message": message},
                timeout=120
            )
            if response.ok:
                data = response.json()
                if data.get("success"):
                    return data.get("response", "")
                else:
                    return f"[ERROR: {data.get('error', 'Unknown error')}]"
            else:
                return f"[HTTP {response.status_code}]"
        except requests.exceptions.Timeout:
            return "[TIMEOUT - External API unavailable]"
        except Exception as e:
            return f"[ERROR: {e}]"

    results = {}

    # Demo 1: What can you help me with?
    print("=" * 60)
    print("DEMO 1: What can you help me with?")
    print("=" * 60)
    response1 = chat("What can you help me with?")
    print(f"Response: {response1[:500]}...")
    # Check if it mentions tasks or available actions
    demo1_pass = any(word in response1.lower() for word in ["task", "add", "list", "help", "manage"])
    results["Demo 1 - Help"] = demo1_pass
    print(f"Result: {'PASS' if demo1_pass else 'FAIL'} (mentions task actions)")

    # Demo 2: Add a task
    print("\n" + "=" * 60)
    print("DEMO 2: Add a task to buy groceries")
    print("=" * 60)
    response2 = chat("Add a task to buy groceries")
    print(f"Response: {response2[:500]}...")
    # Check for confirmation
    demo2_pass = any(word in response2.lower() for word in ["added", "created", "groceries", "task", "id:"])
    results["Demo 2 - Add Task"] = demo2_pass
    print(f"Result: {'PASS' if demo2_pass else 'FAIL'} (task created confirmation)")

    # Demo 3: Show my tasks
    print("\n" + "=" * 60)
    print("DEMO 3: Show my tasks")
    print("=" * 60)
    response3 = chat("Show my tasks")
    print(f"Response: {response3[:500]}...")
    # Check if groceries appears in list
    demo3_pass = "groceries" in response3.lower() or "task" in response3.lower()
    results["Demo 3 - List Tasks"] = demo3_pass
    print(f"Result: {'PASS' if demo3_pass else 'FAIL'} (shows tasks)")

    # Demo 4: Mark task as done
    print("\n" + "=" * 60)
    print("DEMO 4: Mark groceries as done")
    print("=" * 60)
    response4 = chat("Mark groceries as done")
    print(f"Response: {response4[:500]}...")
    # Check for completion confirmation
    demo4_pass = any(word in response4.lower() for word in ["complete", "done", "marked", "✓"])
    results["Demo 4 - Complete Task"] = demo4_pass
    print(f"Result: {'PASS' if demo4_pass else 'FAIL'} (task completed)")

    # Demo 5: Delete task
    print("\n" + "=" * 60)
    print("DEMO 5: Delete the groceries task")
    print("=" * 60)
    response5 = chat("Delete the groceries task")
    print(f"Response: {response5[:500]}...")
    # Check for delete confirmation or prompt
    demo5_pass = any(word in response5.lower() for word in ["delete", "removed", "sure", "confirm", "cannot"])
    results["Demo 5 - Delete Task"] = demo5_pass
    print(f"Result: {'PASS' if demo5_pass else 'FAIL'} (delete handled)")

    # If delete asked for confirmation, confirm it
    if "sure" in response5.lower() or "confirm" in response5.lower():
        print("\n(Confirming delete...)")
        response5b = chat("Yes, delete it")
        print(f"Confirmation response: {response5b[:300]}...")

    # Summary
    print("\n" + "=" * 60)
    print("DEMO SCRIPT SUMMARY")
    print("=" * 60)

    all_passed = True
    for demo, passed in results.items():
        status = "PASS" if passed else "FAIL"
        print(f"  {demo}: {status}")
        if not passed:
            all_passed = False

    passed_count = sum(1 for v in results.values() if v)
    total_count = len(results)

    print(f"\nTotal: {passed_count}/{total_count} demos passed")
    print("\n" + ("All demos passed!" if all_passed else "Some demos need attention"))

    # Note about external API
    if "[TIMEOUT" in str(results) or not all_passed:
        print("\nNote: Some failures may be due to external OpenAI API connectivity.")
        print("The MCP tools and database layer are verified working independently.")

    return 0 if all_passed else 1


if __name__ == "__main__":
    sys.exit(run_demo_script())
