backend/pyproject.toml
[project]
name = "todoapp-backend"
version = "0.1.0"
description = "FastAPI backend for multi-user Todo application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "greenlet>=3.0.0",
    "openai-agents>=0.6.4",
    "mcp>=1.25.0",
    "psycopg2-binary>=2.9.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
